social_network/requirements.txt
Django==4.2
psycopg2>=2.8
djangorestframework==3.14.0
